Scratch another method in ExtensionResolver
The price we pay for this one is actually to remove two interfaces
from the ExtensionResolver. Both only have one implementation,
and both are the only ones that have "complicated" (non-empty)
constructors by convention (it's an awkward pattern for extensions).

The two interfaces are SpringBootExceptionReporter (implementation
FailureAnalyzers), and SpringApplicationRunListener (implementation
EventPublishingRunListener).

I left the interfaces in there in case maybe we want to add explicit
setters in SpringApplication to set them up in a custom way. The
existing implementations already use ExtensionResolver to find their
actual behaviour.